第四次提交

+++ src/main/java/com/example/demo/dao/AbstractBatchDao.java
@@ -11,7 +11,7 @@ public class AbstractBatchDao<T> implements BatchDao<T> {
 
     @Override
     public void batchInsert(List<T> list) {
-        long length = list.size();
+        int length = list.size();
         for (int i = 0; i <length; i++) {
             em.persist(list.get(i));
             if (i % 100 == 0||i==length-1) {//每100条执行一次写入数据库操作
